Set up layout for zone two and created a pit trap
Pit trap will drop character once they walk over it and set up to look for variable of having the rope upgrade. If not will display a UI widget (currently set to main menu but we can change it to something along the lines of "Oh no you fell into a trap! if only you had a rope you could safely climb out!) and then would teleport them back to the raft without their Items. The teleport functions are not currently built into the volume and would need to be as well as the UI itself.